+++ pkg/service/todo_list.go
@@ -13,23 +13,23 @@ func NewTodoListService(repository repository.TodoList) *TodoListService {
 	return &TodoListService{repository: repository}
 }
 
-func (t *TodoListService) Create(userId uint64, list data.List) (uint64, error) {
+func (t *TodoListService) Create(userId uint, list data.List) (uint, error) {
 	return t.repository.Create(userId, list)
 }
 
-func (t *TodoListService) GetAll(userId uint64) ([]data.List, error) {
+func (t *TodoListService) GetAll(userId uint) ([]data.List, error) {
 	return t.repository.GetAll(userId)
 }
 
-func (t *TodoListService) GetByIdAndUserId(userId uint64, listId uint64) (data.List, error) {
+func (t *TodoListService) GetByIdAndUserId(userId uint, listId uint) (data.List, error) {
 	return t.repository.GetByIdAndUserId(userId, listId)
 }
 
-func (t *TodoListService) DeleteByIdAndUserId(userId uint64, listId uint64) error {
+func (t *TodoListService) DeleteByIdAndUserId(userId uint, listId uint) error {
 	return t.repository.DeleteByIdAndUserId(userId, listId)
 }
 
-func (t *TodoListService) UpdateList(userId uint64, listId uint64, updateList data.UpdateListInput) error {
+func (t *TodoListService) UpdateList(userId uint, listId uint, updateList data.UpdateListInput) error {
 	if err := updateList.Validate(); err != nil {
 		return err
 	}
